refactor error boundaries
